Enable SSL, mutual certificate authentication, protected views
