+++ apps/jobs/views.py
@@ -1,14 +1,14 @@
 import logging
 
-from rest_framework import viewsets, mixins, parsers, permissions, status, renderers
+from rest_framework import viewsets, mixins, parsers, status, renderers
 from rest_framework.decorators import action
 from rest_framework.response import Response
 from rest_framework_json_api import parsers as jsapi_parsers
 from influxdb_metrics.loader import log_metric
 
+from apps.authentication import EngineRequest
 from .models import AsyncJob
 from .serializers import AsyncJobSerializer, MessageSerializer
-
 
 LOG = logging.getLogger('transmission')
 
@@ -24,7 +24,7 @@ class JobsViewSet(mixins.ListModelMixin,
     parser_classes = (parsers.JSONParser, jsapi_parsers.JSONParser)
 
     @action(detail=True, methods=['post'],
-            permission_classes=[permissions.AllowAny],
+            permission_classes=[EngineRequest],
             renderer_classes=[renderers.JSONRenderer])
     def message(self, request, version, pk):
         LOG.debug(f'Jobs message called.')
