+++ apps/rpc_client.py
@@ -29,8 +29,12 @@ class RPCError(APIException):
 class RPCClient(object):
     def __init__(self):
         self.url = settings.ENGINE_RPC_URL
-        self.headers = {'content-type': 'application/json'}
         self.payload = {"jsonrpc": "2.0", "id": 0, "params": {}}
+        self.session = requests.session()
+        if settings.ENVIRONMENT in ('PROD', 'STAGE', 'DEV'):
+            self.session.cert = ('/app/client-cert.crt', '/app/client-cert.key')
+            self.session.verify = '/app/ca-bundle.crt'
+        self.session.headers = {'content-type': 'application/json'}
 
     def call(self, method, args=None):
         LOG.debug(f'Calling RPCClient with method {method}.')
@@ -44,8 +48,7 @@ class RPCClient(object):
 
         try:
             with TimingMetric('engine_rpc.call', tags={'method': method}) as timer:
-                response_json = requests.post(self.url, data=json.dumps(self.payload, cls=DecimalEncoder),
-                                              headers=self.headers).json()
+                response_json = self.session.post(self.url, data=json.dumps(self.payload, cls=DecimalEncoder)).json()
                 LOG.info('rpc_client(%s) duration: %.3f', method, timer.elapsed)
 
             if 'error' in response_json:
