+++ compose/nginx/nginx.conf
@@ -16,5 +16,28 @@ server {
         uwsgi_pass django;
         include /etc/nginx/uwsgi_params;
         uwsgi_param HTTP_X_FORWARDED_PROTO https;
+        uwsgi_param X_NGINX_SOURCE alb;
+    }
+}
+
+server {
+    listen 8443 ssl;
+    server_name ~^(.+)$;
+    charset utf-8;
+
+    ssl_certificate     certs/#{DOMAIN}.crt;
+    ssl_certificate_key certs/#{DOMAIN}.key;
+
+    ssl_client_certificate certs/ca-bundle.crt;
+    ssl_verify_client on;
+    ssl_verify_depth 2;
+
+    client_max_body_size 25m;
+
+    location / {
+        uwsgi_pass django;
+        include /etc/nginx/uwsgi_params;
+        uwsgi_param HTTP_X_FORWARDED_PROTO https;
+        uwsgi_param X_NGINX_SOURCE engine;
     }
 }
